File Upload Distance Isochrones for PO Wales
Distance isochrones for post offices at 1 mile, 3 miles and 6 miles thresholds. Using 1202 post offices from Wales + 20km buffer into England. Only 846 post offices in this data set are located in Wales and the rest are in the 20km buffer which falls in England.
According to BEIS, there are 930 post offices in Wales however, by web-scraping the Post Office website in 2019, we found only 846.